include/streamlit_container_requirements.txt
streamlit==1.23.0
langchain
langchain-pinecone
langchain-openai
pinecone-client
openai
python-dotenv